+++ QuickSortInt.h
@@ -5,9 +5,9 @@
 class QuickSortInt
 {
 private:
-	int particionaint(int vetor[], int inicio, int fim);
-	void quicksortint(int vetor[], int ini, int fim);
-	void troca(int &t1, int &t2);
+	int particionaint(long long int vetor[], int inicio, int fim);
+	void quicksortint(long long int vetor[], int ini, int fim);
+	void troca(long long int &t1, long long int &t2);
 	int numTrocas;
 	int numCompar;
 	double tempoGasto;
